Refactor routers into child routers
