+++ server/routes/prizes/index.js
@@ -1,4 +1,4 @@
-// ./routes/prizes.js
+// ./routes/prizes/index.js
 
 const express = require("express");
 const router = express.Router({ mergeParams: true });
