+++ server/routes/stores/prizes.js
@@ -1,0 +1,22 @@
+// ./routes/stores/prizes.js
+
+const express = require("express");
+const router = express.Router({ mergeParams: true });
+
+// ## /{family_id}/stores/{store_id}/prizes/{prize_id}
+// ### GET Method
+// Get details of a reward.
+router.get("/:prize_id", (req, res) => {
+  res.send(`GET /:family_id/stores/:store_id/prizes/:prize_id route hit`);
+});
+
+// ## /{family_id}/stores/{store_id}/prizes/{prize_id}/purchase
+// ### POST Method
+// Purchase a reward from a store.
+router.post("/:prize_id/purchase", (req, res) => {
+  res.send(
+    `POST /:family_id/stores/:store_id/prizes/:prize_id/purchase route hit`
+  );
+});
+
+module.exports = router;
