+++ server/routes/users/index.js
@@ -1,4 +1,4 @@
-// ./routes/users.js
+// ./routes/users/index.js
 
 const express = require("express");
 const router = express.Router({ mergeParams: true });
